knobs: one model, one entry point
Follow-on from the previous commit, which fixed the overlay by giving the two
knob engines the same step floor. Two implementations that agree is still two
places to look and two things to keep in step, and that is exactly how they
drifted apart the first time.

param_pages/movy_knob.mjs is deleted and its model moved into
shared/knob_engine.mjs, which had been the second implementation: a time-based
divisor curve stepping by the module's DECLARED step. The Movy model wins
because it is what the knob grid ships with, so it is the feel already in
people's hands, and because normalising a step to a fraction of the param's
own RANGE is what makes every knob cross in about the same wrist movement
whatever its units. The time-based acceleration is gone; the range
normalisation replaces what it existed to paper over.

There is also only ONE ENTRY POINT. knobStep takes metadata; the config-shaped
knobTick / knobConfigFromMeta pair is deleted rather than aliased, and all
five call sites now pass metadata directly. An alias would have kept every
test green while leaving a second name for the next reader to wire something
up to.

Fixed along the way, all of them consequences of there having been two:

  - shift-fine was float-only in the engine, so shift did nothing at all on an
    int or an enum;
  - in the model, an int under shift moved coarse * 0.1, which for a 1-unit
    step is 0.1 and rounds straight back: an int knob that could not be moved
    AT ALL with shift held, measured at 400,000 detents without crossing
    0..127;
  - two call sites MUTATED chain_params metadata in place to apply a zoom or
    shift step scale, leaking the scaled step to every other reader. They copy
    now;
  - knobConfigFromMeta used to default an absent min/max on the way in. With
    one entry point taking raw metadata, that has to happen in the engine, or
    an undeclared range clamps to NaN and writes the string "NaN" to the
    device.

The FINE rule is written down, because I got it wrong twice while pinning it:
the step becomes a tenth of the coarse step, floored at one whole unit for an
int, AND the detent gate is lifted. The second clause is why shift can make a
gated control four times FASTER while still being the precise gesture -- one
click is one option, instead of four clicks per option and a partial turn
doing nothing.

tests/host/test_knob_engine_single.sh pins the structural half: the deleted
file stays deleted, no aliases come back, all three knob-turning call sites go
through knobStep, and MIN_STEP_RANGE_FRAC is not copied anywhere else.
tests/shadow/test_knob_engine.sh is rewritten from the old divisor-curve
numbers to the properties, including a switch surviving as a one-detent
control and each fine clause.
