+++ src/shared/knob_engine.mjs
@@ -1,175 +1,237 @@
 /**
- * knob_engine.mjs — unified knob acceleration + value-stepping.
- * Port of schwung-rewrite/src/domains/knob_engine.c into JS.
+ * knob_engine.mjs — THE knob model. One implementation, one file.
  *
- * Divisor curve (gap = nowMs - lastTickMs):
- *   gap >  150ms → divisor 16   (fine control)
- *   gap >   50ms → divisor 8
- *   gap == 0 or <= 50ms → divisor 4   (fast sweep)
- *   first tick (lastTickMs == 0) → divisor 1   (intentional "click" on motion start)
+ * Every physical knob turn in the UI comes through here: the Movy knob grid,
+ * the hierarchy list editor and its overlay, the patch editor, the waveform
+ * zoom and marker knobs.
  *
- * Float: step / divisor per tick, where step is max(declared, 1% of range) —
- *        see floatStep(). A declared step is about precision, not sweep, and
- *        taking it literally made a wide-range param need ~80,000 detents.
- * Int:   accumulate ticks; emit ±1 once accum reaches divisor.
- * Enum:  fixed enum_divisor = 10 ticks per option, regardless of count.
- *        Binary toggles and 47-option pickers feel equally snappy.
+ * There used to be two. This file held a time-based divisor curve (16 / 8 / 4
+ * by how fast you turned) stepping by the module's DECLARED step, while
+ * param_pages/movy_knob.mjs held the ported Movy model, which normalises a
+ * step to a fraction of the parameter's own RANGE. A knob therefore behaved
+ * differently depending on which screen you touched it from, reported from
+ * the device as knobs being "super slow in the overlay". The gap was not
+ * subtle:
  *
- * Staleness: gap > 2000ms resets the engine to cold-start (lastTickMs=0),
- * so re-entering an editor after a long pause feels like a fresh edit
- * rather than continuing a stale acceleration curve.
+ *     float 20..20000 step 1      overlay 79,917 detents      grid 200
+ *     int   0..127                overlay    505 detents      grid 127
  *
- * Batched deltas: |direction| > 1 produces proportional motion. The
- * accumulator emits floor(|accum|/divisor) steps per call (int/enum) or
- * step*direction/divisor (float), preserving the old linear "delta * step"
- * behavior under fast jog wheel input.
+ * ...and shift-fine, which is meant to work everywhere, was honoured for
+ * floats in one and ignored for ints and enums, while in the other it could
+ * not move an int at all.
+ *
+ * The Movy model won: it is the one the knob grid ships with, so it is the
+ * feel already in people's hands, and normalising to the range is what makes
+ * every knob cross in about the same wrist movement whatever its units. What
+ * was lost is the time-based acceleration — the range normalisation replaces
+ * what it existed to paper over.
+ *
+ * An intermediate version kept both files, with this one delegating to the
+ * other. That is still two places to look and two things to keep in step, so
+ * the model was moved here and movy_knob.mjs deleted. There is also only one
+ * ENTRY POINT: knobStep, taking metadata. The config-shaped knobTick /
+ * knobConfigFromMeta pair the older call sites used is gone rather than
+ * aliased -- an alias is a second name for the next reader to wire something
+ * up to, which is how the two models drifted apart to begin with.
+ *
+ * ORIGIN: the model below is ported from schwung-movy
+ * (`src/model/{knob-step,store,constants}.ts`, (c) 2026 megadake, MIT —
+ * https://github.com/DimaDake/schwung-movy).
+ *
+ * The core idea: a float/int knob's per-detent step is a fixed FRACTION of
+ * its own range (~1%, MIN_STEP_RANGE_FRAC) rather than the module's declared
+ * `step`, which is a statement about precision and not about sweep. That is
+ * what fixes both a wide-range knob crawling and a narrow one being a hair
+ * trigger.
  */
 
+import { BOOL_OPTION } from "./param_pages/viz.mjs";
+
+/*
+ * The type tags the older call sites build their configs with. They name the
+ * same three kinds the metadata does.
+ */
 export const KNOB_TYPE_FLOAT = "float";
 export const KNOB_TYPE_INT = "int";
 export const KNOB_TYPE_ENUM = "enum";
 
-const KNOB_ACCEL_FAST_MS = 50;
-const KNOB_ACCEL_MED_MS = 150;
-const KNOB_STALE_MS = 2000;   // gap above this → treat as cold start (engine self-resets)
-/* Shared with the Movy knob model (param_pages/movy_knob.mjs), deliberately:
- * the two engines should normalize a wide range to the same fraction even
- * though their acceleration curves differ. */
+/** Physical clicks per value step for a narrow int range, and for every enum. */
+export const ENUM_DELTA_DIV = 4;
+/** Sensitivity multiplier for a continuous arc-rendered knob — every knob in
+ * the Movy layout is one, so this always applies to float/int. */
+export const ARC_DELTA_SCALE = 0.5;
+/** A float/int knob's per-detent step, as a fraction of its own range. */
 export const MIN_STEP_RANGE_FRAC = 0.01;
-/*
- * Shift-held fine adjust, as a divisor of the EFFECTIVE step.
- *
- * It has to live in here, not at the call site. The caller used to hand
- * knobTick a config with its declared step pre-divided by 10, which the
- * range floor above then simply undid -- fine and coarse came out 1.2x
- * apart, i.e. the same. Caught by tests/host/test_param_pages_controller.sh,
- * which had already pinned that fine must be meaningfully finer.
- */
-export const FINE_DIVISOR = 10;
+/** An int range this narrow or narrower is stepped like an enum. */
+export const NARROW_RANGE_MAX = 8;
 
-export function knobInit(initialValue) {
-    return { lastTickMs: 0, value: initialValue, tickAccum: 0 };
+/** schwung-movy model/knob-step.ts detentsPerStep, ported. */
+export function detentsPerStep(meta) {
+    if (meta.type !== "int" || meta.knobAcceleration === "wide") return 1;
+    const range = meta.max - meta.min;
+    return (range >= 2 && range <= NARROW_RANGE_MAX) ? ENUM_DELTA_DIV : 1;
 }
 
-function clampf(v, lo, hi) {
-    if (v < lo) return lo;
-    if (v > hi) return hi;
-    return v;
-}
-
-function tickDivisor(state, nowMs) {
-    if (state.lastTickMs === 0) return 1;
-    const delta = nowMs > state.lastTickMs ? nowMs - state.lastTickMs : 0;
-    /* Stale state — engine self-resets so re-entry feels like a fresh edit. */
-    if (delta > KNOB_STALE_MS) {
-        state.lastTickMs = 0;
-        state.tickAccum = 0;
-        return 1;
-    }
-    if (delta > KNOB_ACCEL_MED_MS) return 16;
-    if (delta > KNOB_ACCEL_FAST_MS) return 8;
-    return 4;
+/** schwung-movy model/knob-step.ts perDetentStep, ported. */
+export function perDetentStep(meta) {
+    const arcScale = ARC_DELTA_SCALE;
+    if (!(meta.max > meta.min)) return (meta.step > 0 ? meta.step : 0.01) * arcScale;
+    const rangeStep = (meta.max - meta.min) * MIN_STEP_RANGE_FRAC;
+    if (meta.type === "int") return Math.round(Math.max(meta.step > 0 ? meta.step : 1, rangeStep) * arcScale);
+    return rangeStep * arcScale;   // float
 }
 
 /**
- * A float knob's per-detent step: the declared one, or 1% of the param's own
- * range, whichever is BIGGER.
- *
- * Reported from the device as knobs being "super slow in the overlay" on the
- * Master FX chain. The declared `step` is a statement about precision, not
- * about sweep, and a module that declares a fine one on a wide range makes
- * this engine unusable: a 20..20000 Hz param with step 1 needed ~80,000
- * detents to cross, against ~200 for the same param on the knob grid.
- *
- * 1% is not a new invention -- MIN_STEP_RANGE_FRAC is the number the Movy
- * knob model already normalizes to, and the grid has shipped with it. This
- * makes the two feel like relatives instead of strangers.
- *
- * A FLOOR, not a replacement: a param whose declared step is already coarser
- * than 1% keeps it. So 0..1 step 0.01 is completely unaffected, which is most
- * of the fleet.
- *
- * The trade is precision at the extreme: on a very wide range you can no
- * longer land on an exact value by turning, and the slow end of the divisor
- * curve (step/16) is the fine control. Typing an exact value is what the
- * editor's edit mode is for.
- *
- * Deliberately NOT applied to the INT path, which is a different bug: that
- * path ignores `config.step` entirely and always moves +/-1 per divisor
- * ticks, so a 0..20000 int is just as slow and cannot be fixed by a floor
- * without also making most values unreachable. Left alone rather than
- * half-fixed.
+ * schwung-movy model/store.ts wideStepCount, ported. Velocity multiplier for
+ * a param that opts into `knobAcceleration: "wide"` — see the module doc:
+ * nothing in the fleet declares this today.
  */
-export function floatStep(config, fine = false) {
-    const declared = config.step > 0 ? config.step : 0.01;
-    const range = (config.max - config.min);
-    const base = (range > 0) ? Math.max(declared, range * MIN_STEP_RANGE_FRAC) : declared;
-    return fine ? base / FINE_DIVISOR : base;
+function wideStepCount(state, direction, nowMs) {
+    let multiplier = 1;
+    const elapsed = state.lastTurnMs > 0 ? nowMs - state.lastTurnMs : Infinity;
+    if (direction === state.lastDirection) {
+        if (elapsed <= 35) multiplier = 250;
+        else if (elapsed <= 90) multiplier = 50;
+        else if (elapsed <= 180) multiplier = 10;
+    }
+    state.lastTurnMs = nowMs;
+    state.lastDirection = direction;
+    return direction * multiplier;
 }
 
-export function knobTick(state, config, direction, nowMs, { fine = false } = {}) {
-    const divisor = tickDivisor(state, nowMs);
-    state.lastTickMs = nowMs;
+/**
+ * A two-state boolean, i.e. exactly what viz.mjs `detectSwitch` draws as a
+ * switch. Kept on the same BOOL_OPTION test so a control cannot be drawn as a
+ * switch but turned like a list (or the reverse).
+ */
+function isSwitchMeta(meta) {
+    const opts = Array.isArray(meta.options) ? meta.options : null;
+    return !!opts && opts.length === 2
+        && opts.every((o) => BOOL_OPTION.test(String(o).trim()));
+}
 
-    if (config.type === KNOB_TYPE_FLOAT) {
-        const step = floatStep(config, fine);
-        const delta = (step / divisor) * direction;
-        state.value = clampf(state.value + delta, config.min, config.max);
-    } else if (config.type === KNOB_TYPE_INT) {
-        /* Accumulator must drain before reversing — eats first N reverse ticks (anti-jitter). */
-        state.tickAccum += direction;
-        const steps = Math.trunc(state.tickAccum / divisor);
-        if (steps !== 0) {
-            state.value = clampf(state.value + steps, config.min, config.max);
-            state.tickAccum -= steps * divisor;
-        }
-    } else if (config.type === KNOB_TYPE_ENUM) {
-        if (!config.enumCount || config.enumCount <= 0) {
-            state.tickAccum = 0;
+/**
+ * Reversing direction drops whatever partial turn was banked the other way.
+ * Without it the residue is spent cancelling the new direction first, so a
+ * reversal costs up to div + (div - 1) detents — 7 at ENUM_DELTA_DIV 4 — and
+ * how many depends on where the previous turn happened to stop, which is why
+ * the same knob feels inconsistent from one reversal to the next.
+ */
+function clearOnReversal(state, delta) {
+    if (state.detentAccum !== 0 && Math.sign(state.detentAccum) !== Math.sign(delta))
+        state.detentAccum = 0;
+}
+
+export function knobInit(initialValue) {
+    return { value: initialValue, detentAccum: 0, lastTurnMs: 0, lastDirection: 0 };
+}
+
+/**
+ * @param {object} state   from movyKnobInit, mutated in place
+ * @param {object} meta    param_meta.mjs metadata (min/max/step/type/options)
+ * @param {number} delta   ±1 per physical detent (this library always decodes
+ *                         one CC message to one detent — see page_input.mjs)
+ * @param {number} nowMs
+ * @param {boolean} [fine] Shift-held precision mode. Not part of Movy's own
+ *                 model (no equivalent was found in its source); grafted on
+ *                 so the gesture works everywhere.
+ *
+ *                 THE RULE, in two clauses, both of which matter:
+ *
+ *                   1. the step becomes a TENTH of the coarse step, floored
+ *                      at one whole unit for an int (a tenth of a 1-unit step
+ *                      is 0.1, which rounds straight back — that was an int
+ *                      knob that could not be moved at all under shift); and
+ *                   2. the detent GATE is lifted, so every detent moves
+ *                      something.
+ *
+ *                 Clause 2 is why shift can make a control FASTER rather than
+ *                 slower, which looks contradictory on a stopwatch: an enum
+ *                 and a narrow int are gated at ENUM_DELTA_DIV detents per
+ *                 option in the coarse feel, so lifting the gate is a
+ *                 fourfold speed-up. It is still precision — with shift held
+ *                 one click is exactly one option, instead of four clicks
+ *                 being one option and a partial turn being nothing at all.
+ *                 That is what you want when placing a value rather than
+ *                 sweeping to one.
+ * @returns {number} the new value
+ */
+export function knobStep(state, meta, delta, nowMs, fine = false) {
+    if (!state || !delta) return state ? state.value : 0;
+
+    /*
+     * Fill in a missing range. Call sites hand us metadata straight off
+     * chain_params, and a module is not obliged to declare min/max -- an
+     * absent one used to arrive as undefined, and the clamp at the bottom
+     * turned the value into NaN, which is then written to the device as the
+     * string "NaN". The old config-shaped entry point defaulted these on the
+     * way in; now that there is only one entry point, it has to do it here.
+     */
+    if (typeof meta.min !== "number" || typeof meta.max !== "number") {
+        const isInt = meta.type === "int";
+        meta = {
+            ...meta,
+            min: typeof meta.min === "number" ? meta.min : 0,
+            max: typeof meta.max === "number" ? meta.max : (isInt ? 127 : 1),
+        };
+    }
+
+    if (meta.kind === "enum" || meta.type === "enum") {
+        /* A switch has two states and no travel, so it flips on ONE detent, in
+         * the direction turned. Running it through the click gate below cost 4
+         * detents to move at all (and up to 7 to come back), which reads as a
+         * dead control: the drawn knob simply does not move when you turn it. */
+        if (isSwitchMeta(meta)) {
+            state.detentAccum = 0;
+            state.value = delta > 0 ? 1 : 0;
             return state.value;
         }
-        /* Fixed ticks-per-option for enums — independent of count so binary
-         * toggles feel as snappy as 47-option pickers. Tunable single number. */
-        const enumDivisor = 10;
-        /* Accumulator must drain before reversing — eats first N reverse ticks (anti-jitter). */
-        state.tickAccum += direction;
-        const steps = Math.trunc(state.tickAccum / enumDivisor);
-        if (steps !== 0) {
-            let iv = Math.round(state.value) + steps;
-            if (iv < 0) iv = 0;
-            if (iv >= config.enumCount) iv = config.enumCount - 1;
-            state.value = iv;
-            state.tickAccum -= steps * enumDivisor;
-        }
+        const div = fine ? 1 : ENUM_DELTA_DIV;
+        clearOnReversal(state, delta);
+        state.detentAccum += delta;
+        const steps = Math.trunc(state.detentAccum / div);
+        if (steps === 0) return state.value;
+        state.detentAccum -= steps * div;
+        const count = Array.isArray(meta.options) ? meta.options.length : (meta.max - meta.min + 1);
+        let iv = Math.round(state.value) + steps;
+        iv = Math.max(0, Math.min(Math.max(0, count - 1), iv));
+        state.value = iv;
+        return state.value;
     }
-    return state.value;
+
+    if (meta.knobAcceleration === "wide") {
+        const scaled = wideStepCount(state, delta > 0 ? 1 : -1, nowMs) * (meta.step > 0 ? meta.step : 1);
+        let next = state.value + scaled;
+        next = Math.max(meta.min, Math.min(meta.max, next));
+        if (meta.type === "int") next = Math.round(next);
+        state.value = next;
+        return next;
+    }
+
+    const div = fine ? 1 : detentsPerStep(meta);
+    let steps = delta;
+    if (div > 1) {
+        clearOnReversal(state, delta);
+        state.detentAccum += delta;
+        steps = Math.trunc(state.detentAccum / div);
+        if (steps === 0) return state.value;
+        state.detentAccum -= steps * div;
+    }
+    /*
+     * FINE is a tenth of the coarse step -- except an int, where a tenth of a
+     * 1-unit step is 0.1 and Math.round below puts it straight back. That was
+     * not "fine is subtle", it was an int knob that could not be moved AT ALL
+     * with shift held: measured at 400,000 detents without crossing 0..127.
+     * An int's finest meaningful move is 1.
+     */
+    const coarseStep = perDetentStep(meta);
+    const stepSize = !fine ? coarseStep
+        : (meta.type === "int" ? Math.max(1, Math.round(coarseStep * 0.1))
+                               : coarseStep * 0.1);
+    let next = state.value + steps * stepSize;
+    next = Math.max(meta.min, Math.min(meta.max, next));
+    if (meta.type === "int") next = Math.round(next);
+    state.value = next;
+    return next;
 }
 
-/* Convert a chain_params metadata entry → KnobConfig accepted by knobTick(). */
-export function knobConfigFromMeta(meta) {
-    if (!meta) return { type: KNOB_TYPE_FLOAT, min: 0, max: 1, step: 0.01 };
-    if (meta.type === "int") {
-        return {
-            type: KNOB_TYPE_INT,
-            min: typeof meta.min === "number" ? meta.min : 0,
-            max: typeof meta.max === "number" ? meta.max : 127,
-            step: meta.step > 0 ? meta.step : 1,
-        };
-    }
-    if (meta.type === "enum") {
-        const opts = Array.isArray(meta.options) ? meta.options : [];
-        return {
-            type: KNOB_TYPE_ENUM,
-            min: 0,
-            max: Math.max(0, opts.length - 1),
-            step: 1,
-            enumCount: opts.length,
-        };
-    }
-    return {
-        type: KNOB_TYPE_FLOAT,
-        min: typeof meta.min === "number" ? meta.min : 0,
-        max: typeof meta.max === "number" ? meta.max : 1,
-        step: meta.step > 0 ? meta.step : 0.01,
-    };
-}
